requirements.txt
python-uinput==0.11.2
pyusb==1.0.2